+++ tests/integration/Validation.JsonStrictValidator.test.js
@@ -50,9 +50,10 @@ describe('Validation.JsonStrictValidator', () => {
   }));
 
   [
-    Version.v1dot2,
+    Version.v1dot5,
+    Version.v1dot4,
     Version.v1dot3,
-    Version.v1dot4
+    Version.v1dot2
   ].forEach((version) => describe(version, () => {
     if (!hasDep) {
       it('throws MissingOptionalDependencyError', async () => {